requirements.txt
insightface>=0.7.3
numpy>=1.22.0
pillow>=9.4.0
pytesseract>=0.3.10
requests>=2.28.0
torch>=1.13.0
tqdm>=4.64.0
transformers>=4.26.0
